chore: move asset service to assets slice and remove direct consumption of it
