--- src/lib/assetService.ts
@@ -1,14 +1,0 @@
-import { AssetService } from '@shapeshiftoss/asset-service'
-
-let service: AssetService | undefined = undefined
-
-export const getAssetService = async () => {
-  if (!service) {
-    service = new AssetService('')
-  }
-  if (!service?.isInitialized) {
-    await service.initialize()
-  }
-
-  return service
-}
